autotests/requirements.txt
pytest
pytest-playwright
allure-pytest
python-dotenv